backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-jose==3.3.0
cryptography==41.0.5
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.1
python-multipart==0.0.6
email-validator==2.1.0
psycopg2-binary==2.9.9
webauthn==1.8.1
python-dotenv==1.0.0
pillow==10.1.0
httpx==0.25.1
aiofiles==23.2.1
pytest==7.4.3
mock==5.1.0
tenacity==8.2.3 